requirements.txt
torch==2.5.1
torchvision
torchaudio
torch_geometric==2.5.1
pandas==2.0.3
tqdm==4.66.2
scikit-learn==1.3.2
matplotlib==3.7.5
seaborn==0.13.2
transformers==4.42.0
pyarrow
fastparquet
tabulate
networkx==3.2
antiberty
# optuna
# scipy==1.13.0

biopython==1.83
pdb2pqr==3.6.2
MDAnalysis==2.7.0